restructure build process
Build process now creates artifacts in an artifacts folder
for now there is one folder for the whole extension
and one for the packaged extension in .vsix format.
Also, the package.json has been split into a vscode specific one
and one that only contains the build scripts.
This way only the files necessary for running the extension are
actually published instead of all files in the repository.
Every file in the extension folder will be part of the published
extension by default.
This also eliminates the need for a .vscodeignore file.
Removed pre* tasks from package.json in favor of &&ing the tasks in
the corresponding tasks as it better conveys the intention.
Also added npm scripts that are suffixed with -only. They are meant to
only run the bare task and none of the task that this task depends on.
This is used for example in CI to run the tasks serially and avoids
running tasks more than once (e.g. test and package depend on build
and build should only be executed once).
Also added convenience task for vscode that builds before debugging
the extension.